+++ registration-processor/post-processor/registration-processor-printing-stage/src/main/java/io/mosip/registration/processor/print/config/PrintStageConfig.java
@@ -3,7 +3,7 @@ package io.mosip.registration.processor.print.config;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import io.mosip.registration.processor.print.stage.PrintStage;
+import io.mosip.registration.processor.print.stage.PrintingStage;
 
 /**
  * @author M1048399
@@ -13,8 +13,8 @@ import io.mosip.registration.processor.print.stage.PrintStage;
 public class PrintStageConfig {
 
 	@Bean 
-	public PrintStage getPrintStage() {
-		return new PrintStage();
+	public PrintingStage getPrintStage() {
+		return new PrintingStage();
 	}
